Organized gear unit tests into multiple files.
